requirements.txt
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
cx_Freeze==7.2.0
filelock==3.15.4
firebase-admin==6.5.0
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.8
msgpack==1.0.8
patchelf==0.17.2.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
pystray==0.19.5
python-xlib==0.33
requests==2.32.3
rsa==4.9
setuptools==70.3.0
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
